New schema file with annotations
This schema file adds the 4 different types of annotations by the
Elizabeth, the PharmD student, as columns (as reflected in the 4
different types of .tsv files shown, which are generated using
annParse.py with the two .ann files shown).

"Insert" statements from a temporary table tmp in the schema - with
annotations.sql file were used so that duplicate keys are retained.