Query accounts for world vista drug classes
View created is a left join of world vista drug interactions with drug group to account for drug classes of drug 1 and/or drug 2 in the interaction.

Then, this view is joined with the NDF-RT interactions to account for all permutations of where drug classes may arise. Query takes a long time to fetch, however.