Results updated, set differences included
When finding set differences, several cases in which the NDF-RT data set had slightly different versions of the same sets of two RxCUI's were found. For instance, some PDDI's in which the same RxCUI was used for drug 1 and drug 2 would vary in "Severity" so that the NDF-RT data recognized these interactions as separate when for the sake of this analysis, it would alter the counts of how many drugs were duplicated in the UNION. Thus, these PDDI's in the NDF-RT data set were consolidated into 1 row rather than 2 or more like they previously were and the query was updated. The result is much lower than before, however, at 8074 overlapping interactions.

The results from queries for the set differences (i.e. PDDI's only found in NDF-RT, PDDI's only found in WorldVista) were also added. These queries are also in overlap-query.sql